practica2/requirements.txt
pycryptodome==3.8.1
requests==2.18.4
termcolor==1.1.0